Pantallas y validaciones de entidades Repuestos, Stocks, Ubicaciones, Sectores, Fabricantes, Marcas, Modelos, UMDNS, Proveedores, Checklists

+++ EquiposTecnicosSN.Web/Controllers/MarcasController.cs
@@ -1,9 +1,10 @@
 ﻿using System.Data.Entity;
-using System.Threading.Tasks;
 using System.Net;
 using System.Web.Mvc;
 using EquiposTecnicosSN.Web.DataContexts;
 using EquiposTecnicosSN.Entities.Equipos.Info;
+using System.Linq;
+using PagedList;
 
 namespace EquiposTecnicosSN.Web.Controllers
 {
@@ -13,32 +14,33 @@ namespace EquiposTecnicosSN.Web.Controllers
         private EquiposDbContext db = new EquiposDbContext();
 
         // GET: Marcas
-        public async Task<ActionResult> Index()
+        public ActionResult Index(int fabricanteId, string searchNombre = null, int page = 1)
         {
-            var marcas = db.Marcas.Include(m => m.Fabricante);
-            return View(await marcas.ToListAsync());
-        }
+            var listPage = db.Marcas
+                .Where(m => m.FabricanteId == fabricanteId)
+                .Where(m => (searchNombre == null || searchNombre == "") || m.Nombre.Contains(searchNombre))
+                .OrderBy(m => m.Nombre)
+                .ToPagedList(page, 10);
 
-        // GET: Marcas/Details/5
-        public async Task<ActionResult> Details(int? id)
-        {
-            if (id == null)
+            if (Request.IsAjaxRequest())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return PartialView("_MarcasList", listPage);
             }
-            Marca marca = await db.Marcas.FindAsync(id);
-            if (marca == null)
-            {
-                return HttpNotFound();
-            }
-            return View(marca);
+
+            ViewBag.FabricanteId = fabricanteId;
+            ViewBag.NombreFabricante = db.Fabricantes.Where(f => f.FabricanteId == fabricanteId).Select(f => f.Nombre).Single();
+            return View(listPage);
         }
 
         // GET: Marcas/Create
-        public ActionResult Create()
+        public ActionResult Create(int fabricanteId)
         {
             ViewBag.FabricanteId = new SelectList(db.Fabricantes, "FabricanteId", "Nombre");
-            return View();
+            var nuevaMarca = new Marca
+            {
+                FabricanteId = fabricanteId
+            };
+            return View(nuevaMarca);
         }
 
         // POST: Marcas/Create
@@ -46,13 +48,13 @@ namespace EquiposTecnicosSN.Web.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include = "MarcaId,Nombre,FabricanteId")] Marca marca)
+        public ActionResult Create(Marca marca)
         {
             if (ModelState.IsValid)
             {
                 db.Marcas.Add(marca);
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                db.SaveChanges();
+                return RedirectToAction("Index", new { fabricanteId = marca.FabricanteId});
             }
 
             ViewBag.FabricanteId = new SelectList(db.Fabricantes, "FabricanteId", "Nombre", marca.FabricanteId);
@@ -60,13 +62,13 @@ namespace EquiposTecnicosSN.Web.Controllers
         }
 
         // GET: Marcas/Edit/5
-        public async Task<ActionResult> Edit(int? id)
+        public ActionResult Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Marca marca = await db.Marcas.FindAsync(id);
+            Marca marca = db.Marcas.Find(id);
             if (marca == null)
             {
                 return HttpNotFound();
@@ -76,16 +78,14 @@ namespace EquiposTecnicosSN.Web.Controllers
         }
 
         // POST: Marcas/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include = "MarcaId,Nombre,FabricanteId")] Marca marca)
+        public ActionResult Edit(Marca marca)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(marca).State = EntityState.Modified;
-                await db.SaveChangesAsync();
+                db.SaveChanges();
                 return RedirectToAction("Index");
             }
             ViewBag.FabricanteId = new SelectList(db.Fabricantes, "FabricanteId", "Nombre", marca.FabricanteId);
@@ -93,13 +93,13 @@ namespace EquiposTecnicosSN.Web.Controllers
         }
 
         // GET: Marcas/Delete/5
-        public async Task<ActionResult> Delete(int? id)
+        public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Marca marca = await db.Marcas.FindAsync(id);
+            Marca marca = db.Marcas.Find(id);
             if (marca == null)
             {
                 return HttpNotFound();
@@ -110,12 +110,37 @@ namespace EquiposTecnicosSN.Web.Controllers
         // POST: Marcas/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int id)
         {
-            Marca marca = await db.Marcas.FindAsync(id);
-            db.Marcas.Remove(marca);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
+            var equiposCount = db.Equipos
+                .Where(e => e.InformacionHardware.MarcaId == id)
+                .Count();
+
+            var modelosCount = db.Modelos
+                .Where(m => m.MarcaId == id)
+                .Count();
+
+            var marca = db.Marcas.Find(id);
+
+            if (equiposCount == 0 && modelosCount == 0)
+            {
+                ViewBag.FabricanteId = marca.FabricanteId;
+                db.Marcas.Remove(marca);
+                db.SaveChanges();
+                return RedirectToAction("Index", new { fabricanteId = ViewBag.FabricanteId });
+            }
+
+            if (equiposCount > 0)
+            {
+                ModelState.AddModelError("", "La Marca no puede eliminarse ya que posee equipos asociados.");
+            }
+
+            if (modelosCount > 0)
+            {
+                ModelState.AddModelError("", "La Marca no puede eliminarse ya que posee Modelos asociados.");
+            }
+
+            return View(marca);
         }
 
         protected override void Dispose(bool disposing)
